* pr/18552:
  Polish "Remove boxing"
  Remove boxing

Closes gh-18552

+++ spring-boot-project/spring-boot/src/test/java/org/springframework/boot/jdbc/metadata/CommonsDbcp2DataSourcePoolMetadataTests.java
@@ -61,7 +61,7 @@ class CommonsDbcp2DataSourcePoolMetadataTests
 	@Test
 	void getPoolUsageWithUnlimitedPool() {
 		DataSourcePoolMetadata unlimitedDataSource = createDataSourceMetadata(0, -1);
-		assertThat(unlimitedDataSource.getUsage()).isEqualTo(Float.valueOf(-1F));
+		assertThat(unlimitedDataSource.getUsage()).isEqualTo(-1f);
 	}
 
 	@Override
